Removed some Actor C++ classes Added Leather Boots w/ Localization Support Updated readme
