+++ Content/Localization/Game/Game.csv
@@ -42,3 +42,4 @@ Date/Time,Word Count,en,pt-PT
 2022.07.26-14.28.30,102,102,81
 2022.07.26-14.40.14,103,103,81
 2022.07.26-15.19.28,105,105,82
+2022.07.30-16.28.10,111,111,84
