+++ Source/TestProject/Private/Item/FItemStructure.cpp
@@ -28,3 +28,34 @@ FItemStructure::FItemStructure()
 	Defense = 0;
 	Health = 0;
 }
+
+FItemStructure::FItemStructure(FName ID, UTexture2D* Icon, FText Name, FName Description, EItemType ItemType, EItemQuality Quality, uint8 PriceValue,
+	bool IsStackable, uint8 MaxStackSize, bool IsDroppable, USkeletalMesh* SkeletalMesh, UStaticMesh* WorldMesh, float Duration, 
+	EEquipmentSlot EquipmentSlot, uint8 CurrentLevel, uint8 MaxLevel, uint8 AmountToAddPerLevel, uint8 Strength, uint8 Endurance, uint8 Dexterity, uint8 Intelligence, 
+	uint8 Attack, uint8 Defense, uint8 Health)
+{
+	ID = ID;
+	Icon = Icon;
+	Name = Name;
+	Description = Description;
+	ItemType = ItemType;
+	Quality = Quality;
+	PriceValue = PriceValue;
+	IsStackable = IsStackable;
+	MaxStackSize = MaxStackSize;
+	IsDroppable = IsDroppable;
+	SkeletalMesh = SkeletalMesh;
+	WorldMesh = WorldMesh;
+	Duration = Duration;
+	EquipmentSlot = EquipmentSlot;
+	CurrentLevel = CurrentLevel;
+	MaxLevel = MaxLevel;
+	AmountToAddPerLevel = AmountToAddPerLevel;
+	Strength = Strength;
+	Endurance = Endurance;
+	Dexterity = Dexterity;
+	Intelligence = Intelligence;
+	Attack = Attack;
+	Defense = Defense;
+	Health = Health;
+}
