--- Source/TestProject/Public/Actors/BoneSword.h
@@ -1,18 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "WorldActor.h"
-#include "BoneSword.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class INVENTORYSYSTEMCPP_API ABoneSword : public AWorldActor
-{
-	GENERATED_BODY()
-
-	ABoneSword();
-};
