+++ Source/TestProject/Public/Actors/SimpleAxe.h
@@ -1,0 +1,18 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "WorldActor.h"
+#include "SimpleAxe.generated.h"
+
+/**
+ * 
+ */
+UCLASS()
+class INVENTORYSYSTEMCPP_API ASimpleAxe : public AWorldActor
+{
+	GENERATED_BODY()
+	
+	ASimpleAxe();
+};
