--- Source/TestProject/Public/SimpleAxe.h
@@ -1,18 +1,0 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "WorldActor.h"
-#include "SimpleAxe.generated.h"
-
-/**
- * 
- */
-UCLASS()
-class INVENTORYSYSTEMCPP_API ASimpleAxe : public AWorldActor
-{
-	GENERATED_BODY()
-	
-	ASimpleAxe();
-};
